fix(interpolate): introduce deprecation warning for interpolation
